Main svc new public events (#19)
* Текущее состояние ветки

* Наработки по ветке

* Изменения и дополнения

* Текущее состояние ветки

* Дополнил метод getEvent теперь он заполняет все поля сущности event

* Добавил создание списка Event по списку id Исправил ошибку в вызове clientController.getHits

* Исправления по промежуточным результатам тестов

* Добавляются свежие правки. Доступен профиль work для запуска всех модулей с базой h2 для работы в среде IDEA

* Еще свежие правки.

* Настройка работы с сервером статистики.

* Исправлены баги. Пока отключил логи sql-запросов в профиле work

* Обновлены мапперы для соответствия данных запросу

* Сделал валидацию параметров метода контроллера, добавил тест

* модификаторы в TakeHitsDto

* fix getEvent

* fix style

* fix style

* fix test

* fix events null

---------

+++ ewm-service/src/main/java/ru/practicum/users/service/ParticipationRequestService.java
@@ -36,7 +36,7 @@ public class ParticipationRequestService {
                 .toList();
     }
 
-    public long getConfirmedRequests(long eventId) {
+    public int getConfirmedRequests(long eventId) {
         return requestRepository
                 .countConfirmedRequestsByStatusAndEventId(ParticipationRequestStatus.CONFIRMED, eventId);
     }
